move Intervals strategy to common
