+++ hypothesis-python/tests/cover/test_intervalset.py
@@ -13,29 +13,7 @@ import pytest
 from hypothesis import HealthCheck, assume, example, given, settings, strategies as st
 from hypothesis.internal.intervalsets import IntervalSet
 
-
-def build_intervals(ls):
-    ls.sort()
-    result = []
-    for u, l in ls:
-        v = u + l
-        if result:
-            a, b = result[-1]
-            if u <= b + 1:
-                result[-1] = (a, v)
-                continue
-        result.append((u, v))
-    return result
-
-
-def IntervalLists(min_size=0):
-    return st.lists(
-        st.tuples(st.integers(0, 200), st.integers(0, 20)),
-        min_size=min_size,
-    ).map(build_intervals)
-
-
-Intervals = st.builds(IntervalSet, IntervalLists())
+from tests.common.strategies import IntervalLists, Intervals
 
 
 @given(Intervals)
